requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
scikit-learn>=1.5.0
pandas>=2.1.0
numpy>=1.26.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
